% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_capital_juros_composto.R
\name{calcular_capital_juros_composto}
\alias{calcular_capital_juros_composto}
\title{Calcular capital de juros compostos}
\usage{
calcular_capital_juros_composto(montante, juros, tempo)
}
\arguments{
\item{montante}{numero do montante}

\item{juros}{numero decimal dos juros}

\item{tempo}{numero do tempo}
}
\value{
um numero
}
\description{
Calcular capital de juros compostos
}
\examples{
calcular_capital_juros_composto(2200,0.1,12)

}
