% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_capital_juros_simples.R
\name{calcular_capital_juros_simples}
\alias{calcular_capital_juros_simples}
\title{Calcular capital de juros simples}
\usage{
calcular_capital_juros_simples(montante, juros, tempo)
}
\arguments{
\item{montante}{numero do montante}

\item{juros}{numero decimal dos juros}

\item{tempo}{numero do tempo}
}
\value{
um numero
}
\description{
Calcular capital de juros simples
}
\examples{
calcular_capital_juros_simples(2200,0.1,12)

}
