% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_juros_composto.R
\name{calcular_juros_composto}
\alias{calcular_juros_composto}
\title{Calcular juros compostos}
\usage{
calcular_juros_composto(capital, taxa, tempo)
}
\arguments{
\item{capital}{numero}

\item{taxa}{numero}

\item{tempo}{numero}
}
\value{
numero
}
\description{
Calcular juros compostos
}
\examples{
calcular_juros_composto(1000,0.1,12)

}
