% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_montante_juros_compostos.R
\name{calcular_montante_juros_composto}
\alias{calcular_montante_juros_composto}
\title{Calcular montante de juros compostos}
\usage{
calcular_montante_juros_composto(capital, juros, tempo)
}
\arguments{
\item{capital}{número inicial do capital}

\item{juros}{número decimal dos juros}

\item{tempo}{número do tempo}
}
\value{
um número
}
\description{
Calcular montante de juros compostos
}
\examples{
calcular_montante_juros_composto(1000,0.1,12)

}
