% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_taxa_juros_composto.R
\name{calcular_taxa_juros_composto}
\alias{calcular_taxa_juros_composto}
\title{Calcular taxa de juros simples}
\usage{
calcular_taxa_juros_composto(capital, montante, tempo)
}
\arguments{
\item{capital}{numero}

\item{montante}{numero}

\item{tempo}{numero}
}
\value{
numero
}
\description{
Calcular taxa de juros simples
}
\examples{
calcular_taxa_juros_composto(1000,2200,12)

}
