% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_taxa_juros_simples.R
\name{calcular_taxa_juros_simples}
\alias{calcular_taxa_juros_simples}
\title{Calcular taxa de juros simples}
\usage{
calcular_taxa_juros_simples(capital, juros, tempo)
}
\arguments{
\item{capital}{número}

\item{tempo}{número}

\item{montante}{número}
}
\value{
número
}
\description{
Calcular taxa de juros simples
}
\examples{
calcular_taxa_juros_simples(1000,2200,12)

}
