% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_tempo_juros_composto.R
\name{calcular_tempo_juros_composto}
\alias{calcular_tempo_juros_composto}
\title{Calcular tempo de juros compostos}
\usage{
calcular_tempo_juros_composto(capital, montante, juros)
}
\arguments{
\item{capital}{número}

\item{montante}{número}

\item{juros}{número}
}
\value{
número
}
\description{
Calcular tempo de juros compostos
}
\examples{
calcular_tempo_juros_composto(1000,3138.428,0.1)

}
