% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_tempo_juros_simples.R
\name{calcular_tempo_juros_simples}
\alias{calcular_tempo_juros_simples}
\title{Calcular tempo de juros simples}
\usage{
calcular_tempo_juros_simples(capital, montante, juros)
}
\arguments{
\item{capital}{numero}

\item{montante}{numero}

\item{juros}{numero}
}
\value{
numero
}
\description{
Calcular tempo de juros simples
}
\examples{
calcular_tempo_juros_simples(1000,2200,0.1)

}
