% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_juros_simples.R
\name{calcular_juros_simples}
\alias{calcular_juros_simples}
\title{Calcular juros simples}
\usage{
calcular_juros_simples(capital, taxa, tempo)
}
\arguments{
\item{capital}{número}

\item{taxa}{número}

\item{tempo}{número}
}
\value{
número
}
\description{
Calcular juros simples
}
\examples{
calcular_juros_simples(1000,0.1,12)

}
