% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_montante_juros_simples.R
\name{calcular_montante_juros_simples}
\alias{calcular_montante_juros_simples}
\title{Calcular montante de juros simples}
\usage{
calcular_montante_juros_simples(capital, juros, tempo)
}
\arguments{
\item{capital}{número inicial do capital}

\item{juros}{número decimal dos juros}

\item{tempo}{número do tempo}
}
\value{
um número
}
\description{
Essa função é usada para calcular o montante baseado no capital inicial, juros e tempo de rendimento.
}
\examples{
calcular_montante_juros_simples(1000,0.1,12)

}
